[Baekjoon>greedy] BJ_13305 - 성공

+++ Baekjoon/greedy/BJ_13305_greedy.java
@@ -11,8 +11,8 @@ public class BJ_13305_greedy {
 
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int n = Integer.parseInt(br.readLine());
-		int[] dist = new int[n - 1];
-		int[] value = new int[n];
+		long[] dist = new long[n - 1];
+		long[] value = new long[n];
 		StringTokenizer st = new StringTokenizer(br.readLine());
 		for (int i = 0; i < n - 1; i++) {
 			dist[i] = Integer.parseInt(st.nextToken());
@@ -25,9 +25,9 @@ public class BJ_13305_greedy {
 		System.out.println(solution(n, dist, value));
 	}
 
-	public static int solution(int n, int[] dist, int[] value) {
-		int curValue = value[0];
-		int sum = curValue * dist[0];
+	public static long solution(int n, long[] dist, long[] value) {
+		long curValue = value[0];
+		long sum = curValue * dist[0];
 		for (int i = 1; i < n - 1; i++) {
 			if (curValue < value[i]) {
 				sum += curValue * dist[i];
